feat(swiftui home): extract EndOfFeedScrollView from CardCollection, for reuse
